"Fix typo in PaymentForm address field, log items in onFinish, add order state to store"
This commit fixes a typo in the PaymentForm component where the "adress" field name is corrected to "address". It also modifies the onFinish function in PaymentForm to log the items in the cart. Additionally, the order state is added to the store by importing the order reducer in the store file.