"Rename fetchItemSuccess action and related selectors in OrderSlice"
Explanation: The diff shows that the 'fetchItemSuccess' action and related selectors in the OrderSlice have been renamed to 'fetchOrderSuccess' and 'selectItem' respectively. The 'fetchItemSuccess' action was used to update the 'item' state property with the payload received from the server. However, it seems that the state property has been renamed to 'order' in this commit, hence the need to rename the action and selectors accordingly.